% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyGrouping.R
\name{ApplyGrouping}
\alias{ApplyGrouping}
\title{ApplyGrouping}
\usage{
ApplyGrouping(
  data,
  originGrouping,
  from = "FullRegionOfOrigin",
  to = "GroupedRegionOfOrigin",
  asFactor = FALSE
)
}
\arguments{
\item{data}{Data. Required.}

\item{originGrouping}{List object with mapping. Required.}

\item{from}{Column name for mapping from}

\item{to}{Column name for mapping to}

\item{asFactor}{Logical indicating to convert "from" and "to" columns to factor.}
}
\value{
inputData
}
\description{
Applies grouping map to the input data
}
\examples{
\dontrun{
ApplyGrouping(
  inputData,
  originGrouping,
  from = 'FullRegionOfOrigin',
  to = 'GroupedRegionOfOrigin'
)
}

}
