% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AppManager.R, R/Task.R
\name{AppManager}
\alias{AppManager}
\title{AppManager}
\description{
R6 class for representing the app manaager

R6 class for representing a background task
}
\examples{
pppManager <- AppManager$new()
\dontrun{
tast <- Task$new()
}
}
