% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineData.R
\name{CombineData}
\alias{CombineData}
\title{CombineData}
\usage{
CombineData(set1, set2)
}
\arguments{
\item{set1}{Set 1, case-based data. Optional.}

\item{set2}{Set 2, aggregated data. Optional.}
}
\value{
Combined data set - a list of aggregated data sets
}
\description{
Combines two aggregated data sets for HIV Model.
Data is set 2 overrides data in set 1.
}
\examples{
\dontrun{
CombineData(set1, set2)
}

}
