% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAvailableStrata.R
\name{GetAvailableStrata}
\alias{GetAvailableStrata}
\title{GetAvailableStrata}
\usage{
GetAvailableStrata(dt, colNames)
}
\arguments{
\item{dt}{dt}

\item{colNames}{colNames}
}
\value{
list
}
\description{
Get available strata
}
