% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAverageHIVModelOutputs.R
\name{GetAverageHIVModelOutputs}
\alias{GetAverageHIVModelOutputs}
\title{GetAverageHIVModelOutputs}
\usage{
GetAverageHIVModelOutputs(hivModels)
}
\arguments{
\item{hivModels}{hivModels}
}
\value{
data.table
}
\description{
GetAverageHIVModelOutputs
}
