% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAveragePreMigrCounts.R
\name{GetAveragePreMigrCounts}
\alias{GetAveragePreMigrCounts}
\title{GetAveragePreMigrCounts}
\usage{
GetAveragePreMigrCounts(caseData = NULL, migrConnFlag, dataAfterMigr)
}
\arguments{
\item{caseData}{caseData}

\item{migrConnFlag}{migrConnFlag}

\item{dataAfterMigr}{dataAfterMigr}
}
\value{
list
}
\description{
GetAveragePreMigrCounts
}
