% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCuts.R
\name{GetCuts}
\alias{GetCuts}
\title{GetCuts}
\usage{
GetCuts(data, count = 3L)
}
\arguments{
\item{data}{data}

\item{count}{count of cuts}
}
\value{
vector of integers
}
\description{
Get cuts for VarX to be used in RD estimation with time trend
}
\examples{
data <- data.table::data.table(
  VarXs = 58L:68L
)
GetCuts(data, count = 3L)

}
