% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHIVPlotData.R
\name{GetHIVPlotData}
\alias{GetHIVPlotData}
\title{GetHIVPlotData}
\usage{
GetHIVPlotData(mainFitOutputs, bootstrapFitStats = NULL, parameters = NULL)
}
\arguments{
\item{mainFitOutputs}{mainFitOutputs}

\item{bootstrapFitStats}{bootstrapFitStats}

\item{parameters}{parameters}
}
\value{
data.table
}
\description{
Get HIV plot data
}
\examples{
\dontrun{
GetHIVPlotData(
  mainFitOutputs,
  bootstrapFitStats,
  parameters,
  migrConnFlag
)
}

}
