% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMigrantConfBounds.R
\name{GetMigrantConfBounds}
\alias{GetMigrantConfBounds}
\title{GetMigrantConfBounds}
\usage{
GetMigrantConfBounds(
  data,
  strat = c(),
  region = "ALL",
  regionColumn = "MigrantRegionOfOrigin",
  minPresentRatio = 0.9,
  detailed = FALSE
)
}
\arguments{
\item{data}{data}

\item{strat}{strat}

\item{region}{region}

\item{regionColumn}{regionColumn}

\item{minPresentRatio}{minPresentRatio}

\item{detailed}{detailed}
}
\value{
object (data.table or list)
}
\description{
Compute confidence bounds for migrant related statistics.
}
\examples{
\dontrun{
GetMigrantConfBounds(
  data,
  strat = c(),
  region = 'ALL',
  regionColumn = 'MigrantRegionOfOrigin',
  minPresentRatio = 0.9,
  detailed = FALSE
)
}

}
