% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMigrantOutputStats.R
\name{GetMigrantOutputStats}
\alias{GetMigrantOutputStats}
\title{GetMigrantOutputStats}
\usage{
GetMigrantOutputStats(data)
}
\arguments{
\item{data}{data}
}
\value{
list
}
\description{
Get statistics for migrant module output
}
\examples{
\dontrun{
GetMigrantOutputStats(data)
}

}
