% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOriginGroupingPreset.R
\name{GetOriginGroupingPreset}
\alias{GetOriginGroupingPreset}
\title{GetOriginGroupingPreset}
\usage{
GetOriginGroupingPreset(preset = "REPCOUNTRY + UNK + OTHER", distr = NULL)
}
\arguments{
\item{preset}{Grouping preset type. Default = 'REPCOUNTRY + UNK + OTHER'}

\item{distr}{Distribution of RegionOfOrigin. Default = NULL}
}
\description{
Get mapping from RegionOfOrigin to GroupOfOrigin
}
\examples{
distr <- data.table::data.table(
  FullRegionOfOrigin = c('REPCOUNTRY', 'SUBAFR'),
  Count = c(1536, 2237)
)
GetOriginGroupingPreset('REPCOUNTRY + UNK + 3 most prevalent regions + OTHER', distr)

}
