% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPreMigrCounts.R
\name{GetPreMigrCounts}
\alias{GetPreMigrCounts}
\title{GetPreMigrCounts}
\usage{
GetPreMigrCounts(caseData = NULL)
}
\arguments{
\item{caseData}{caseData}
}
\value{
list
}
\description{
GetPreMigrCounts
}
