% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HIVModelDataManager.R
\name{HIVModelDataManager}
\alias{HIVModelDataManager}
\title{HIVModelDataManager}
\description{
R6 class for representing the HIV Model manaager
}
\examples{
hivModelMgr <- HIVModelDataManager$new()
}
