% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostProcessModelOutputs.R
\name{PostProcessModelOutputs}
\alias{PostProcessModelOutputs}
\title{PostProcessModelOutputs}
\usage{
PostProcessModelOutputs(
  modelOutputs,
  preMigrCounts,
  migrConnFlag,
  dataAfterMigr
)
}
\arguments{
\item{modelOutputs}{modelOutputs}

\item{preMigrCounts}{preMigrCounts}

\item{migrConnFlag}{migrConnFlag}

\item{dataAfterMigr}{dataAfterMigr}
}
\value{
data.table
}
\description{
PostProcessModelOutputs
}
