% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareMigrantData.R
\name{PrepareMigrantData}
\alias{PrepareMigrantData}
\title{PrepareMigrantData}
\usage{
PrepareMigrantData(data, seed = NULL)
}
\arguments{
\item{data}{data}

\item{seed}{sees}
}
\value{
list
}
\description{
Pre-processes data as input for migrant module
}
\examples{
\dontrun{
PrepareMigrantData(data, seed)
}

}
