% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Messages.R
\name{PrintStopHeader}
\alias{PrintStopHeader}
\title{PrintStopHeader}
\usage{
PrintStopHeader(startTime = NULL, stopTime = NULL)
}
\arguments{
\item{startTime}{Start time as POSIXct}

\item{stopTime}{End time as POSIXct}
}
\description{
PrintStopHeader
}
\examples{
PrintStopHeader()

}
