% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunApp.R
\name{RunApp}
\alias{RunApp}
\title{RunApp}
\usage{
RunApp(port = NULL, launchBrowser = FALSE, stopOnSessionEnded = FALSE)
}
\arguments{
\item{port}{Port}

\item{launchBrowser}{launchBrowser}

\item{stopOnSessionEnded}{stopOnSessionEnded}
}
\value{
NULL (invisibly)
}
\description{
Run the application.
}
\examples{
\dontrun{
RunApp()
}

}
