% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertStringToDate.R
\name{ConvertStringToDate}
\alias{ConvertStringToDate}
\title{ConvertStringToDate}
\usage{
ConvertStringToDate(str)
}
\arguments{
\item{str}{str}
}
\value{
Date
}
\description{
ConvertStringToDate
}
\examples{
ConvertStringToDate(str = c(
  NA, 'UNK', '', '2020', '2020-Q1', '2020-Q02', '2020-W01', '2020-W45', '2020-05-20', '2020-5-20'
))

}
