% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertStringToDate.R
\name{ConvertStringToDate}
\alias{ConvertStringToDate}
\title{ConvertStringToDate}
\usage{
ConvertStringToDate(str)
}
\arguments{
\item{str}{str}
}
\value{
Date
}
\description{
ConvertStringToDate
}
\examples{
ConvertStringToDate(c(NA, '2020', '2020-05-20'))

}
