% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAveragePreMigrCounts.R
\name{GetAveragePreMigrCounts}
\alias{GetAveragePreMigrCounts}
\title{GetAveragePreMigrCounts}
\usage{
GetAveragePreMigrCounts(caseData = NULL)
}
\arguments{
\item{caseData}{caseData}
}
\value{
list
}
\description{
GetAveragePreMigrCounts
}
