% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPreMigrCounts.R
\name{GetPreMigrCounts}
\alias{GetPreMigrCounts}
\title{GetPreMigrCounts}
\usage{
GetPreMigrCounts(caseData = NULL, migrConnFlag, dataAfterMigr)
}
\arguments{
\item{caseData}{caseData}

\item{migrConnFlag}{migrConnFlag}

\item{dataAfterMigr}{dataAfterMigr}
}
\value{
list
}
\description{
GetPreMigrCounts
}
